+++ src/endpoints/git.rs
@@ -1,0 +1,317 @@
+// REPO OPERATIONS
+
+use std::path::{Components, PathBuf};
+use std::sync::atomic::Ordering;
+use git2::{Repository, StatusOptions};
+use rocket::{get, State};
+use rocket::http::{ContentType, Status};
+use rocket::response::status;
+use crate::static_vars::NET_IS_ENABLED;
+use crate::structs::{AppSettings, GitStatusRecord};
+use crate::utils::json_responses::{make_bad_json_data_response, make_good_json_data_response};
+use crate::utils::paths::{check_path_components, os_slash_str};
+
+#[get("/list-local-repos")]
+pub(crate) fn list_local_repos(state: &State<AppSettings>) -> status::Custom<(ContentType, String)> {
+    let root_path = state.repo_dir.lock().unwrap().clone();
+    let server_paths = std::fs::read_dir(root_path).unwrap();
+    let mut repos: Vec<String> = Vec::new();
+    for server_path in server_paths {
+        let uw_server_path_ob = server_path.unwrap().path();
+        let uw_server_path_ob2 = uw_server_path_ob.clone();
+        let server_leaf = uw_server_path_ob2.file_name().unwrap();
+        for org_path in std::fs::read_dir(uw_server_path_ob).unwrap() {
+            let uw_org_path_ob = org_path.unwrap().path();
+            let uw_org_path_ob2 = uw_org_path_ob.clone();
+            let org_leaf = uw_org_path_ob2.file_name().unwrap();
+            for repo_path in std::fs::read_dir(uw_org_path_ob).unwrap() {
+                let uw_repo_path_ob = repo_path.unwrap().path();
+                let repo_leaf = uw_repo_path_ob.file_name().unwrap();
+                let repo_url_string = format!(
+                    "{}/{}/{}",
+                    server_leaf.to_str().unwrap(),
+                    org_leaf.to_str().unwrap(),
+                    repo_leaf.to_str().unwrap()
+                );
+                repos.push(repo_url_string);
+            }
+        }
+    }
+    let quoted_repos: Vec<String> = repos
+        .into_iter()
+        .map(|str: String| format!("{}", str))
+        .collect();
+    status::Custom(
+        Status::Ok,
+        (
+            ContentType::JSON,
+            serde_json::to_string(&quoted_repos).unwrap(),
+        ),
+    )
+}
+
+#[get("/add-and-commit/<repo_path..>")]
+pub(crate) async fn add_and_commit(
+    state: &State<AppSettings>,
+    repo_path: PathBuf,
+) -> status::Custom<(ContentType, String)> {
+    let repo_path_string: String = state.repo_dir.lock().unwrap().clone()
+        + os_slash_str()
+        + &repo_path.display().to_string().clone();
+    let result = match Repository::open(repo_path_string) {
+        Ok(repo) => {
+            repo.index()
+                .unwrap()
+                .add_all(&["."], git2::IndexAddOption::DEFAULT, None)
+                .unwrap();
+            repo.index().unwrap().write().unwrap();
+            let mut index = repo.index().unwrap();
+            let oid = index.write_tree().unwrap();
+            let signature = repo.signature().unwrap();
+            let parent_commit = repo.head().unwrap().peel_to_commit().unwrap();
+            let tree = repo.find_tree(oid).unwrap();
+            repo.commit(
+                Some("HEAD"),
+                &signature,
+                &signature,
+                "Updated by Pithekos",
+                &tree,
+                &[&parent_commit],
+            )
+                .unwrap();
+            status::Custom(
+                Status::Ok,
+                (
+                    ContentType::JSON,
+                    make_good_json_data_response("ok".to_string()),
+                ),
+            )
+        }
+        Err(e) => status::Custom(
+            Status::InternalServerError,
+            (
+                ContentType::JSON,
+                make_bad_json_data_response(
+                    format!("could not add/commit repo: {}", e).to_string(),
+                ),
+            ),
+        ),
+    };
+    result
+}
+#[get("/fetch-repo/<repo_path..>")]
+pub(crate) async fn fetch_repo(
+    state: &State<AppSettings>,
+    repo_path: PathBuf,
+) -> status::Custom<(ContentType, String)> {
+    if !NET_IS_ENABLED.load(Ordering::Relaxed) {
+        return status::Custom(
+            Status::Unauthorized,
+            (
+                ContentType::JSON,
+                make_bad_json_data_response("offline mode".to_string()),
+            ),
+        );
+    }
+    let mut path_components: Components<'_> = repo_path.components();
+    if check_path_components(&mut path_components.clone()) {
+        let source = path_components
+            .next()
+            .unwrap()
+            .as_os_str()
+            .to_str()
+            .unwrap();
+        let org = path_components
+            .next()
+            .unwrap()
+            .as_os_str()
+            .to_str()
+            .unwrap();
+        let mut repo = path_components
+            .next()
+            .unwrap()
+            .as_os_str()
+            .to_str()
+            .unwrap()
+            .to_string();
+        if repo.ends_with(".git") {
+            let repo_vec = repo.split(".").collect::<Vec<&str>>();
+            let short_repo = &repo_vec[0..repo_vec.len() - 1];
+            let short_repo_string = short_repo.join("/");
+            repo = short_repo_string.as_str().to_owned();
+        }
+        let url = "https://".to_string() + &repo_path.display().to_string().replace("\\", "/");
+        match Repository::clone(
+            &url,
+            state.repo_dir.lock().unwrap().clone()
+                + os_slash_str()
+                + source
+                + os_slash_str()
+                + org
+                + os_slash_str()
+                + repo.as_str(),
+        ) {
+            Ok(_repo) => status::Custom(
+                Status::Ok,
+                (
+                    ContentType::JSON,
+                    make_good_json_data_response("ok".to_string()),
+                ),
+            ),
+            Err(e) => {
+                println!("Error:{}", e);
+                return status::Custom(
+                    Status::BadRequest,
+                    (
+                        ContentType::JSON,
+                        make_bad_json_data_response(
+                            format!("could not clone repo: {}", e).to_string(),
+                        ),
+                    ),
+                );
+            }
+        }
+    } else {
+        status::Custom(
+            Status::BadRequest,
+            (
+                ContentType::JSON,
+                make_bad_json_data_response("bad repo path".to_string()),
+            ),
+        )
+    }
+}
+
+#[get("/delete-repo/<repo_path..>")]
+pub(crate) async fn delete_repo(
+    state: &State<AppSettings>,
+    repo_path: PathBuf,
+) -> status::Custom<(ContentType, String)> {
+    let path_components: Components<'_> = repo_path.components();
+    if check_path_components(&mut path_components.clone()) {
+        let path_to_delete = state.repo_dir.lock().unwrap().clone()
+            + os_slash_str()
+            + &repo_path.display().to_string();
+        match std::fs::remove_dir_all(path_to_delete) {
+            Ok(_) => status::Custom(
+                Status::Ok,
+                (
+                    ContentType::JSON,
+                    make_good_json_data_response("ok".to_string()),
+                ),
+            ),
+            Err(e) => status::Custom(
+                Status::BadRequest,
+                (
+                    ContentType::JSON,
+                    make_bad_json_data_response(
+                        format!("could not delete repo: {}", e).to_string(),
+                    ),
+                ),
+            ),
+        }
+    } else {
+        status::Custom(
+            Status::BadRequest,
+            (
+                ContentType::JSON,
+                make_bad_json_data_response("bad repo path".to_string()),
+            ),
+        )
+    }
+}
+
+#[get("/status/<repo_path..>")]
+pub(crate) async fn git_status(
+    state: &State<AppSettings>,
+    repo_path: PathBuf,
+) -> status::Custom<(ContentType, String)> {
+    let repo_path_string: String = state.repo_dir.lock().unwrap().clone()
+        + os_slash_str()
+        + &repo_path.display().to_string().clone();
+    match Repository::open(repo_path_string) {
+        Ok(repo) => {
+            if repo.is_bare() {
+                return status::Custom(
+                    Status::BadRequest,
+                    (
+                        ContentType::JSON,
+                        make_bad_json_data_response("cannot get status of bare repo".to_string()),
+                    ),
+                );
+            };
+            let mut opts = StatusOptions::new();
+            opts.include_untracked(true);
+            match repo.statuses(Some(&mut opts)) {
+                Ok(statuses) => {
+                    let mut status_changes = Vec::new();
+                    for entry in statuses
+                        .iter()
+                        .filter(|e| e.status() != git2::Status::CURRENT)
+                    {
+                        let i_status = match entry.status() {
+                            s if s.contains(git2::Status::INDEX_NEW)
+                                || s.contains(git2::Status::WT_NEW) =>
+                                {
+                                    "new"
+                                }
+                            s if s.contains(git2::Status::INDEX_MODIFIED)
+                                || s.contains(git2::Status::WT_MODIFIED) =>
+                                {
+                                    "modified"
+                                }
+                            s if s.contains(git2::Status::INDEX_DELETED)
+                                || s.contains(git2::Status::WT_DELETED) =>
+                                {
+                                    "deleted"
+                                }
+                            s if s.contains(git2::Status::INDEX_RENAMED)
+                                || s.contains(git2::Status::WT_RENAMED) =>
+                                {
+                                    "renamed"
+                                }
+                            s if s.contains(git2::Status::INDEX_TYPECHANGE)
+                                || s.contains(git2::Status::WT_TYPECHANGE) =>
+                                {
+                                    "type_change"
+                                }
+                            _ => "",
+                        };
+
+                        if entry.status().contains(git2::Status::IGNORED) {
+                            continue;
+                        }
+                        status_changes.push(GitStatusRecord {
+                            path: entry.path().unwrap().to_string(),
+                            change_type: i_status.to_string(),
+                        });
+                        // println!("{} ({})", entry.path().unwrap(), istatus);
+                    }
+                    status::Custom(
+                        Status::Ok,
+                        (
+                            ContentType::JSON,
+                            serde_json::to_string_pretty(&status_changes).unwrap(),
+                        ),
+                    )
+                }
+                Err(e) => status::Custom(
+                    Status::InternalServerError,
+                    (
+                        ContentType::JSON,
+                        make_bad_json_data_response(
+                            format!("could not get repo status: {}", e).to_string(),
+                        ),
+                    ),
+                ),
+            }
+        }
+        Err(e) => status::Custom(
+            Status::InternalServerError,
+            (
+                ContentType::JSON,
+                make_bad_json_data_response(format!("could not open repo: {}", e).to_string()),
+            ),
+        ),
+    }
+}
