+++ src/endpoints/mod.rs
@@ -2,3 +2,10 @@ pub(crate) mod settings;
 pub(crate) mod net;
 pub(crate) mod debug;
 pub(crate) mod i18n;
+pub(crate) mod sse;
+pub(crate) mod navigation;
+pub(crate) mod gitea;
+pub(crate) mod git;
+pub(crate) mod burrito;
+pub(crate) mod clients;
+pub(crate) mod error;
