+++ src/lib.rs
@@ -3,1065 +3,31 @@
 mod tests;
 
 use copy_dir::copy_dir;
-use git2::{Repository, StatusOptions};
-use hallomai::transform;
-use rocket::form::Form;
 use rocket::fs::{relative, FileServer};
-use rocket::http::{ContentType, Status};
-use rocket::response::{status, stream, Redirect};
-use rocket::tokio::time;
-use rocket::{catch, catchers, get, post, routes, Build, Request, Rocket, State};
+use rocket::{catchers, routes, Build, Rocket};
 use serde_json::{json, Map, Value};
 use std::collections::{BTreeMap, VecDeque};
 use std::io::Write;
-use std::path::{Components, Path, PathBuf};
+use std::path::{Path};
 use std::process::exit;
-use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::{Arc, Mutex};
-use std::time::Duration;
 use std::{env, fs};
-use ureq;
-use uuid::Uuid;
 mod structs;
 use crate::structs::{
-    Bcv,
-    AuthRequest,
     AppSettings,
-    RemoteRepoRecord,
-    GitStatusRecord,
-    MetadataSummary,
-    Upload,
     Client,
-    PublicClient,
-    ContentOrRedirect
+    PublicClient
 };
 mod utils;
-use crate::utils::json_responses::{
-    make_good_json_data_response,
-    make_bad_json_data_response,
-};
 use crate::utils::paths::{
     os_slash_str,
     maybe_os_quoted_path_str,
-    check_path_components,
-    check_path_string_components,
     home_dir_string
 };
-use crate::utils::mime::mime_types;
 mod endpoints;
 
-// CONSTANTS AND STATE
-
-static NET_IS_ENABLED: AtomicBool = AtomicBool::new(false);
-static DEBUG_IS_ENABLED: AtomicBool = AtomicBool::new(false);
-
-// SSE
-#[get("/")]
-async fn notifications_stream<'a>(
-    msgs: &'a State<MsgQueue>,
-    state: &'a State<AppSettings>,
-) -> stream::EventStream![stream::Event + 'a] {
-    stream::EventStream! {
-        let mut count = 0;
-        let mut interval = time::interval(Duration::from_millis(500));
-        yield stream::Event::retry(Duration::from_secs(1));
-        loop {
-            while !msgs.lock().unwrap().is_empty() {
-                let msg = msgs.lock().unwrap().pop_front().unwrap();
-                yield stream::Event::data(msg)
-                    .event("misc")
-                    .id(format!("{}", count));
-                count+=1;
-                interval.tick().await;
-            };
-            yield stream::Event::data(
-                    match NET_IS_ENABLED.load(Ordering::Relaxed) {
-                        true => "enabled",
-                        false => "disabled"
-                    }
-            )
-            .event("net_status")
-            .id(format!("{}", count));
-            count+=1;
-            yield stream::Event::data(
-                    match DEBUG_IS_ENABLED.load(Ordering::Relaxed) {
-                        true => "enabled",
-                        false => "disabled"
-                    }
-            )
-            .event("debug")
-            .id(format!("{}", count));
-            count+=1;
-            let bcv = state.bcv.lock().unwrap().clone();
-            yield stream::Event::data(
-                format!("{}--{}--{}", bcv.book_code, bcv.chapter, bcv.verse)
-            )
-            .event("bcv")
-            .id(format!("{}", count));
-            count+=1;
-            let typography = state.typography.lock().unwrap().clone();
-            yield stream::Event::data(
-                format!("{}--{}--{}", typography.font_set, typography.size, typography.direction)
-            )
-            .event("typography")
-            .id(format!("{}", count));
-            count+=1;
-            let gitea_endpoints = state.gitea_endpoints.clone();
-            let auth_tokens = state.auth_tokens.lock().unwrap().clone();
-            for (ep_name, ep_endpoint) in gitea_endpoints {
-                yield stream::Event::data(
-                    format!("{}--{}--{}", ep_name, ep_endpoint, auth_tokens.contains_key(&ep_name))
-                )
-                .event("auth")
-                .id(format!("{}", count));
-                count+=1;
-            }
-            interval.tick().await;
-        }
-    }
-}
-
-// NAVIGATION
-#[get("/bcv")]
-fn get_bcv(state: &State<AppSettings>) -> status::Custom<(ContentType, String)> {
-    let bcv = state.bcv.lock().unwrap().clone();
-    match serde_json::to_string(&bcv) {
-        Ok(v) => status::Custom(Status::Ok, (ContentType::JSON, v)),
-        Err(e) => status::Custom(
-            Status::InternalServerError,
-            (
-                ContentType::JSON,
-                make_bad_json_data_response(format!(
-                    "Could not parse bcv state as JSON object: {}",
-                    e
-                )),
-            ),
-        ),
-    }
-}
-
-#[post("/bcv/<book_code>/<chapter>/<verse>")]
-fn post_bcv(
-    state: &State<AppSettings>,
-    book_code: &str,
-    chapter: u16,
-    verse: u16,
-) -> status::Custom<(ContentType, String)> {
-    *state.bcv.lock().unwrap() = Bcv {
-        book_code: book_code.to_string(),
-        chapter,
-        verse,
-    };
-    status::Custom(
-        Status::Ok,
-        (
-            ContentType::JSON,
-            make_good_json_data_response("ok".to_string()),
-        ),
-    )
-}
-
-// GITEA
-
-#[get("/endpoints")]
-fn get_gitea_endpoints(state: &State<AppSettings>) -> status::Custom<(ContentType, String)> {
-    status::Custom(
-        Status::Ok,
-        (ContentType::JSON, serde_json::to_string(&state.gitea_endpoints).unwrap()),
-    )
-}
-
-#[get("/remote-repos/<gitea_server>/<gitea_org>")]
-fn gitea_remote_repos(
-    gitea_server: &str,
-    gitea_org: &str,
-) -> status::Custom<(ContentType, String)> {
-    if !NET_IS_ENABLED.load(Ordering::Relaxed) {
-        return status::Custom(
-            Status::Unauthorized,
-            (
-                ContentType::JSON,
-                make_bad_json_data_response("offline mode".to_string()),
-            ),
-        );
-    }
-    let gitea_path = format!("https://{}/api/v1/orgs/{}/repos", gitea_server, gitea_org);
-    match ureq::get(gitea_path.as_str()).call() {
-        Ok(r) => match r.into_json::<Value>() {
-            Ok(j) => {
-                let mut records: Vec<RemoteRepoRecord> = Vec::new();
-                for json_record in j.as_array().unwrap() {
-                    let latest = &json_record["catalog"]["latest"];
-                    records.push(RemoteRepoRecord {
-                        name: json_record["name"].as_str().unwrap().to_string(),
-                        abbreviation: json_record["abbreviation"].as_str().unwrap().to_string(),
-                        description: json_record["description"].as_str().unwrap().to_string(),
-                        avatar_url: json_record["avatar_url"].as_str().unwrap().to_string(),
-                        flavor: json_record["flavor"].as_str().unwrap().to_string(),
-                        flavor_type: json_record["flavor_type"].as_str().unwrap().to_string(),
-                        language_code: json_record["language"].as_str().unwrap().to_string(),
-                        script_direction: json_record["language_direction"]
-                            .as_str()
-                            .unwrap()
-                            .to_string(),
-                        branch_or_tag: match latest["branch_or_tag_name"].as_str() {
-                            Some(s) => s.to_string(),
-                            _ => "".to_string(),
-                        },
-                        clone_url: match latest["released"].as_str() {
-                            Some(s) => s.to_string(),
-                            _ => "".to_string(),
-                        },
-                    });
-                }
-                status::Custom(
-                    Status::Ok,
-                    (ContentType::JSON, serde_json::to_string(&records).unwrap()),
-                )
-            }
-            Err(e) => {
-                return status::Custom(
-                    Status::InternalServerError,
-                    (
-                        ContentType::JSON,
-                        make_bad_json_data_response(format!(
-                            "could not serve GITEA server response as JSON string: {}",
-                            e
-                        )),
-                    ),
-                )
-            }
-        },
-        Err(e) => status::Custom(
-            Status::BadGateway,
-            (
-                ContentType::JSON,
-                make_bad_json_data_response(
-                    format!("could not read from GITEA server: {}", e).to_string(),
-                ),
-            ),
-        ),
-    }
-}
-
-#[get("/login/<token_key>/<redir_path..>")]
-fn gitea_proxy_login(
-    state: &State<AppSettings>,
-    token_key: String,
-    redir_path: PathBuf,
-) -> ContentOrRedirect {
-    if !NET_IS_ENABLED.load(Ordering::Relaxed) {
-        return ContentOrRedirect::Content(
-            status::Custom(
-                Status::Unauthorized,
-                (
-                    ContentType::JSON,
-                    make_bad_json_data_response("offline mode".to_string()),
-                ),
-            )
-        );
-    }
-    if !state.gitea_endpoints.contains_key(&token_key) {
-        return ContentOrRedirect::Content(
-            status::Custom(
-                Status::BadRequest,
-                (
-                    ContentType::JSON,
-                    make_bad_json_data_response(format!(
-                        "Unknown GITEA endpoint name: {}",
-                        token_key
-                    )),
-                ),
-            )
-        );
-    }
-    // Remove any existing token
-    state
-        .auth_tokens
-        .lock()
-        .unwrap()
-        .remove(&token_key);
-    // Store request info
-    let code = Uuid::new_v4().to_string();
-    let mut auth_requests = state
-        .auth_requests
-        .lock()
-        .unwrap();
-    auth_requests.remove(&token_key);
-    auth_requests.insert(
-        token_key.clone(),
-        AuthRequest {
-            code: code.clone(),
-            redirect_uri: redir_path.display().to_string(),
-            timestamp: std::time::SystemTime::now()
-        }
-    );
-    // Do redirect
-    ContentOrRedirect::Redirect(
-        Redirect::to(
-            format!("{}/auth?client_code={}", state.gitea_endpoints[&token_key].clone(), &code)
-        )
-    )
-}
-
-#[get("/logout/<token_key>")]
-fn gitea_proxy_logout(
-    state: &State<AppSettings>,
-    token_key: String
-) -> ContentOrRedirect {
-    if !NET_IS_ENABLED.load(Ordering::Relaxed) {
-        return ContentOrRedirect::Content(
-            status::Custom(
-                Status::Unauthorized,
-                (
-                    ContentType::JSON,
-                    make_bad_json_data_response("offline mode".to_string()),
-                ),
-            )
-        );
-    }
-    if !state.gitea_endpoints.contains_key(&token_key) {
-        return ContentOrRedirect::Content(
-            status::Custom(
-                Status::BadRequest,
-                (
-                    ContentType::JSON,
-                    make_bad_json_data_response(format!(
-                        "Unknown GITEA endpoint name: {}",
-                        token_key
-                    )),
-                ),
-            )
-        );
-    }
-    // Logout of proxy server
-    let logout_url = format!("{}/logout", state.gitea_endpoints[&token_key].clone());
-    println!("{}", logout_url);
-    match ureq::get(logout_url.as_str()).call() {
-        Ok(_) => {
-            // Remove any existing token
-            state
-                .auth_tokens
-                .lock()
-                .unwrap()
-                .remove(&token_key);
-            // Do redirect
-            ContentOrRedirect::Redirect(
-                Redirect::to("/")
-            )
-        },
-        Err(e) => ContentOrRedirect::Content(
-            status::Custom(
-                Status::InternalServerError,
-                (
-                    ContentType::JSON,
-                    make_bad_json_data_response(format!(
-                        "Error on logout from proxy {}: {}",
-                        token_key,
-                        e
-                    )),
-                ),
-            )
-        )
-    }
-}
-
-// REPO OPERATIONS
-
-#[get("/list-local-repos")]
-fn list_local_repos(state: &State<AppSettings>) -> status::Custom<(ContentType, String)> {
-    let root_path = state.repo_dir.lock().unwrap().clone();
-    let server_paths = fs::read_dir(root_path).unwrap();
-    let mut repos: Vec<String> = Vec::new();
-    for server_path in server_paths {
-        let uw_server_path_ob = server_path.unwrap().path();
-        let uw_server_path_ob2 = uw_server_path_ob.clone();
-        let server_leaf = uw_server_path_ob2.file_name().unwrap();
-        for org_path in fs::read_dir(uw_server_path_ob).unwrap() {
-            let uw_org_path_ob = org_path.unwrap().path();
-            let uw_org_path_ob2 = uw_org_path_ob.clone();
-            let org_leaf = uw_org_path_ob2.file_name().unwrap();
-            for repo_path in fs::read_dir(uw_org_path_ob).unwrap() {
-                let uw_repo_path_ob = repo_path.unwrap().path();
-                let repo_leaf = uw_repo_path_ob.file_name().unwrap();
-                let repo_url_string = format!(
-                    "{}/{}/{}",
-                    server_leaf.to_str().unwrap(),
-                    org_leaf.to_str().unwrap(),
-                    repo_leaf.to_str().unwrap()
-                );
-                repos.push(repo_url_string);
-            }
-        }
-    }
-    let quoted_repos: Vec<String> = repos
-        .into_iter()
-        .map(|str: String| format!("{}", str))
-        .collect();
-    status::Custom(
-        Status::Ok,
-        (
-            ContentType::JSON,
-            serde_json::to_string(&quoted_repos).unwrap(),
-        ),
-    )
-}
-
-#[get("/add-and-commit/<repo_path..>")]
-async fn add_and_commit(
-    state: &State<AppSettings>,
-    repo_path: PathBuf,
-) -> status::Custom<(ContentType, String)> {
-    let repo_path_string: String = state.repo_dir.lock().unwrap().clone()
-        + os_slash_str()
-        + &repo_path.display().to_string().clone();
-    let result = match Repository::open(repo_path_string) {
-        Ok(repo) => {
-            repo.index()
-                .unwrap()
-                .add_all(&["."], git2::IndexAddOption::DEFAULT, None)
-                .unwrap();
-            repo.index().unwrap().write().unwrap();
-            let mut index = repo.index().unwrap();
-            let oid = index.write_tree().unwrap();
-            let signature = repo.signature().unwrap();
-            let parent_commit = repo.head().unwrap().peel_to_commit().unwrap();
-            let tree = repo.find_tree(oid).unwrap();
-            repo.commit(
-                Some("HEAD"),
-                &signature,
-                &signature,
-                "Updated by Pithekos",
-                &tree,
-                &[&parent_commit],
-            )
-                .unwrap();
-            status::Custom(
-                Status::Ok,
-                (
-                    ContentType::JSON,
-                    make_good_json_data_response("ok".to_string()),
-                ),
-            )
-        }
-        Err(e) => status::Custom(
-            Status::InternalServerError,
-            (
-                ContentType::JSON,
-                make_bad_json_data_response(
-                    format!("could not add/commit repo: {}", e).to_string(),
-                ),
-            ),
-        ),
-    };
-    result
-}
-#[get("/fetch-repo/<repo_path..>")]
-async fn fetch_repo(
-    state: &State<AppSettings>,
-    repo_path: PathBuf,
-) -> status::Custom<(ContentType, String)> {
-    if !NET_IS_ENABLED.load(Ordering::Relaxed) {
-        return status::Custom(
-            Status::Unauthorized,
-            (
-                ContentType::JSON,
-                make_bad_json_data_response("offline mode".to_string()),
-            ),
-        );
-    }
-    let mut path_components: Components<'_> = repo_path.components();
-    if check_path_components(&mut path_components.clone()) {
-        let source = path_components
-            .next()
-            .unwrap()
-            .as_os_str()
-            .to_str()
-            .unwrap();
-        let org = path_components
-            .next()
-            .unwrap()
-            .as_os_str()
-            .to_str()
-            .unwrap();
-        let mut repo = path_components
-            .next()
-            .unwrap()
-            .as_os_str()
-            .to_str()
-            .unwrap()
-            .to_string();
-        if repo.ends_with(".git") {
-            let repo_vec = repo.split(".").collect::<Vec<&str>>();
-            let short_repo = &repo_vec[0..repo_vec.len() - 1];
-            let short_repo_string = short_repo.join("/");
-            repo = short_repo_string.as_str().to_owned();
-        }
-        let url = "https://".to_string() + &repo_path.display().to_string().replace("\\", "/");
-        match Repository::clone(
-            &url,
-            state.repo_dir.lock().unwrap().clone()
-                + os_slash_str()
-                + source
-                + os_slash_str()
-                + org
-                + os_slash_str()
-                + repo.as_str(),
-        ) {
-            Ok(_repo) => status::Custom(
-                Status::Ok,
-                (
-                    ContentType::JSON,
-                    make_good_json_data_response("ok".to_string()),
-                ),
-            ),
-            Err(e) => {
-                println!("Error:{}", e);
-                return status::Custom(
-                    Status::BadRequest,
-                    (
-                        ContentType::JSON,
-                        make_bad_json_data_response(
-                            format!("could not clone repo: {}", e).to_string(),
-                        ),
-                    ),
-                );
-            }
-        }
-    } else {
-        status::Custom(
-            Status::BadRequest,
-            (
-                ContentType::JSON,
-                make_bad_json_data_response("bad repo path".to_string()),
-            ),
-        )
-    }
-}
-
-#[get("/delete-repo/<repo_path..>")]
-async fn delete_repo(
-    state: &State<AppSettings>,
-    repo_path: PathBuf,
-) -> status::Custom<(ContentType, String)> {
-    let path_components: Components<'_> = repo_path.components();
-    if check_path_components(&mut path_components.clone()) {
-        let path_to_delete = state.repo_dir.lock().unwrap().clone()
-            + os_slash_str()
-            + &repo_path.display().to_string();
-        match fs::remove_dir_all(path_to_delete) {
-            Ok(_) => status::Custom(
-                Status::Ok,
-                (
-                    ContentType::JSON,
-                    make_good_json_data_response("ok".to_string()),
-                ),
-            ),
-            Err(e) => status::Custom(
-                Status::BadRequest,
-                (
-                    ContentType::JSON,
-                    make_bad_json_data_response(
-                        format!("could not delete repo: {}", e).to_string(),
-                    ),
-                ),
-            ),
-        }
-    } else {
-        status::Custom(
-            Status::BadRequest,
-            (
-                ContentType::JSON,
-                make_bad_json_data_response("bad repo path".to_string()),
-            ),
-        )
-    }
-}
-
-#[get("/status/<repo_path..>")]
-async fn git_status(
-    state: &State<AppSettings>,
-    repo_path: PathBuf,
-) -> status::Custom<(ContentType, String)> {
-    let repo_path_string: String = state.repo_dir.lock().unwrap().clone()
-        + os_slash_str()
-        + &repo_path.display().to_string().clone();
-    match Repository::open(repo_path_string) {
-        Ok(repo) => {
-            if repo.is_bare() {
-                return status::Custom(
-                    Status::BadRequest,
-                    (
-                        ContentType::JSON,
-                        make_bad_json_data_response("cannot get status of bare repo".to_string()),
-                    ),
-                );
-            };
-            let mut opts = StatusOptions::new();
-            opts.include_untracked(true);
-            match repo.statuses(Some(&mut opts)) {
-                Ok(statuses) => {
-                    let mut status_changes = Vec::new();
-                    for entry in statuses
-                        .iter()
-                        .filter(|e| e.status() != git2::Status::CURRENT)
-                    {
-                        let i_status = match entry.status() {
-                            s if s.contains(git2::Status::INDEX_NEW)
-                                || s.contains(git2::Status::WT_NEW) =>
-                                {
-                                    "new"
-                                }
-                            s if s.contains(git2::Status::INDEX_MODIFIED)
-                                || s.contains(git2::Status::WT_MODIFIED) =>
-                                {
-                                    "modified"
-                                }
-                            s if s.contains(git2::Status::INDEX_DELETED)
-                                || s.contains(git2::Status::WT_DELETED) =>
-                                {
-                                    "deleted"
-                                }
-                            s if s.contains(git2::Status::INDEX_RENAMED)
-                                || s.contains(git2::Status::WT_RENAMED) =>
-                                {
-                                    "renamed"
-                                }
-                            s if s.contains(git2::Status::INDEX_TYPECHANGE)
-                                || s.contains(git2::Status::WT_TYPECHANGE) =>
-                                {
-                                    "type_change"
-                                }
-                            _ => "",
-                        };
-
-                        if entry.status().contains(git2::Status::IGNORED) {
-                            continue;
-                        }
-                        status_changes.push(GitStatusRecord {
-                            path: entry.path().unwrap().to_string(),
-                            change_type: i_status.to_string(),
-                        });
-                        // println!("{} ({})", entry.path().unwrap(), istatus);
-                    }
-                    status::Custom(
-                        Status::Ok,
-                        (
-                            ContentType::JSON,
-                            serde_json::to_string_pretty(&status_changes).unwrap(),
-                        ),
-                    )
-                }
-                Err(e) => status::Custom(
-                    Status::InternalServerError,
-                    (
-                        ContentType::JSON,
-                        make_bad_json_data_response(
-                            format!("could not get repo status: {}", e).to_string(),
-                        ),
-                    ),
-                ),
-            }
-        }
-        Err(e) => status::Custom(
-            Status::InternalServerError,
-            (
-                ContentType::JSON,
-                make_bad_json_data_response(format!("could not open repo: {}", e).to_string()),
-            ),
-        ),
-    }
-}
-
-// METADATA OPERATIONS
-#[get("/metadata/raw/<repo_path..>")]
-async fn raw_metadata(
-    state: &State<AppSettings>,
-    repo_path: PathBuf,
-) -> status::Custom<(ContentType, String)> {
-    let path_components: Components<'_> = repo_path.components();
-    if check_path_components(&mut path_components.clone()) {
-        let path_to_serve = state.repo_dir.lock().unwrap().clone()
-            + os_slash_str()
-            + &repo_path.display().to_string()
-            + "/metadata.json";
-        match fs::read_to_string(path_to_serve) {
-            Ok(v) => status::Custom(Status::Ok, (ContentType::JSON, v)),
-            Err(e) => status::Custom(
-                Status::BadRequest,
-                (
-                    ContentType::JSON,
-                    make_bad_json_data_response(
-                        format!("could not read metadata: {}", e).to_string(),
-                    ),
-                ),
-            ),
-        }
-    } else {
-        status::Custom(
-            Status::BadRequest,
-            (
-                ContentType::JSON,
-                make_bad_json_data_response("bad repo path".to_string()),
-            ),
-        )
-    }
-}
-
-#[get("/metadata/summary/<repo_path..>")]
-async fn summary_metadata(
-    state: &State<AppSettings>,
-    repo_path: PathBuf,
-) -> status::Custom<(ContentType, String)> {
-    let path_components: Components<'_> = repo_path.components();
-    if check_path_components(&mut path_components.clone()) {
-        let path_to_serve = state.repo_dir.lock().unwrap().clone()
-            + os_slash_str()
-            + &repo_path.display().to_string()
-            + os_slash_str()
-            + "metadata.json";
-        println!("{}", path_to_serve);
-        let file_string = match fs::read_to_string(path_to_serve) {
-            Ok(v) => v,
-            Err(e) => {
-                return status::Custom(
-                    Status::BadRequest,
-                    (
-                        ContentType::JSON,
-                        make_bad_json_data_response(
-                            format!("could not read metadata: {}", e).to_string(),
-                        ),
-                    ),
-                )
-            }
-        };
-        let raw_metadata_struct: Value =
-            match serde_json::from_str(file_string.as_str()) {
-                Ok(v) => v,
-                Err(e) => {
-                    return status::Custom(
-                        Status::BadRequest,
-                        (
-                            ContentType::JSON,
-                            make_bad_json_data_response(
-                                format!("could not parse metadata: {}", e).to_string(),
-                            ),
-                        ),
-                    )
-                }
-            };
-        let summary = MetadataSummary {
-            name: raw_metadata_struct["identification"]["name"]["en"]
-                .as_str()
-                .unwrap()
-                .to_string(),
-            description: match raw_metadata_struct["identification"]["description"]["en"].clone() {
-                Value::String(v) => v.as_str().to_string(),
-                Value::Null => "".to_string(),
-                _ => "?".to_string(),
-            },
-            flavor_type: raw_metadata_struct["type"]["flavorType"]["name"]
-                .as_str()
-                .unwrap()
-                .to_string(),
-            flavor: raw_metadata_struct["type"]["flavorType"]["flavor"]["name"]
-                .as_str()
-                .unwrap()
-                .to_string(),
-            language_code: raw_metadata_struct["languages"][0]["tag"]
-                .as_str()
-                .unwrap()
-                .to_string(),
-            script_direction: match raw_metadata_struct["languages"][0]["scriptDirection"].clone() {
-                Value::String(v) => v.as_str().to_string(),
-                _ => "?".to_string(),
-            },
-        };
-        match serde_json::to_string(&summary) {
-            Ok(v) => status::Custom(Status::Ok, (ContentType::JSON, v)),
-            Err(e) => status::Custom(
-                Status::InternalServerError,
-                (
-                    ContentType::JSON,
-                    make_bad_json_data_response(
-                        format!("could not serialize metadata: {}", e).to_string(),
-                    ),
-                ),
-            ),
-        }
-    } else {
-        status::Custom(
-            Status::BadRequest,
-            (
-                ContentType::JSON,
-                make_bad_json_data_response("bad repo path!".to_string()),
-            ),
-        )
-    }
-}
-
-// INGREDIENT OPERATIONS
-
-#[get("/ingredient/raw/<repo_path..>?<ipath>")]
-async fn raw_ingredient(
-    state: &State<AppSettings>,
-    repo_path: PathBuf,
-    ipath: String,
-) -> status::Custom<(ContentType, String)> {
-    let path_components: Components<'_> = repo_path.components();
-    if check_path_components(&mut path_components.clone())
-        && check_path_string_components(ipath.clone())
-    {
-        let path_to_serve = state.repo_dir.lock().unwrap().clone()
-            + os_slash_str()
-            + &repo_path.display().to_string()
-            + "/ingredients/"
-            + ipath.as_str();
-        match fs::read_to_string(path_to_serve) {
-            Ok(v) => {
-                let mut split_ipath = ipath.split(".").clone();
-                let mut suffix = "unknown";
-                if let Some(_) = split_ipath.next() {
-                    if let Some(second) = split_ipath.next() {
-                        suffix = second;
-                    }
-                }
-                status::Custom(
-                    Status::Ok,
-                    (
-                        match mime_types().get(suffix) {
-                            Some(t) => t.clone(),
-                            None => ContentType::new("application", "octet-stream"),
-                        },
-                        v,
-                    ),
-                )
-            }
-            Err(e) => status::Custom(
-                Status::BadRequest,
-                (
-                    ContentType::JSON,
-                    make_bad_json_data_response(
-                        format!("could not read ingredient content: {}", e).to_string(),
-                    ),
-                ),
-            ),
-        }
-    } else {
-        status::Custom(
-            Status::BadRequest,
-            (
-                ContentType::JSON,
-                make_bad_json_data_response("bad repo path".to_string()),
-            ),
-        )
-    }
-}
-
-#[get("/ingredient/as-usj/<repo_path..>?<ipath>")]
-async fn get_ingredient_as_usj(
-    state: &State<AppSettings>,
-    repo_path: PathBuf,
-    ipath: String,
-) -> status::Custom<(ContentType, String)> {
-    let path_components: Components<'_> = repo_path.components();
-    if check_path_components(&mut path_components.clone())
-        && check_path_string_components(ipath.clone())
-    {
-        let path_to_serve = state.repo_dir.lock().unwrap().clone()
-            + os_slash_str()
-            + &repo_path.display().to_string()
-            + "/ingredients/"
-            + ipath.as_str();
-        match fs::read_to_string(path_to_serve) {
-            Ok(v) => status::Custom(
-                Status::Ok,
-                (
-                    ContentType::JSON,
-                    transform(v, "usfm".to_string(), "usj".to_string()),
-                ),
-            ),
-            Err(e) => status::Custom(
-                Status::BadRequest,
-                (
-                    ContentType::JSON,
-                    make_bad_json_data_response(
-                        format!("could not read ingredient content: {}", e).to_string(),
-                    ),
-                ),
-            ),
-        }
-    } else {
-        status::Custom(
-            Status::BadRequest,
-            (
-                ContentType::JSON,
-                make_bad_json_data_response("bad repo path".to_string()),
-            ),
-        )
-    }
-}
-
-#[post(
-    "/ingredient/as-usj/<repo_path..>?<ipath>",
-    format = "multipart/form-data",
-    data = "<form>"
-)]
-async fn post_ingredient_as_usj(
-    state: &State<AppSettings>,
-    repo_path: PathBuf,
-    ipath: String,
-    mut form: Form<Upload<'_>>,
-) -> status::Custom<(ContentType, String)> {
-    let path_components: Components<'_> = repo_path.components();
-    let destination = state.repo_dir.lock().unwrap().clone()
-        + os_slash_str()
-        + &repo_path.display().to_string()
-        + "/ingredients/"
-        + ipath.clone().as_str();
-    if check_path_components(&mut path_components.clone())
-        && check_path_string_components(ipath)
-        && fs::metadata(destination.clone()).is_ok()
-    {
-        let _ = form
-            .file
-            .persist_to(transform(
-                destination,
-                "usj".to_string(),
-                "usfm".to_string(),
-            ))
-            .await;
-        status::Custom(
-            Status::Ok,
-            (
-                ContentType::JSON,
-                make_good_json_data_response("ok".to_string()),
-            ),
-        )
-    } else {
-        status::Custom(
-            Status::BadRequest,
-            (
-                ContentType::JSON,
-                make_bad_json_data_response("bad repo path".to_string()),
-            ),
-        )
-    }
-}
-
-#[get("/ingredient/prettified/<repo_path..>?<ipath>")]
-async fn get_ingredient_prettified(
-    state: &State<AppSettings>,
-    repo_path: PathBuf,
-    ipath: String,
-) -> status::Custom<(ContentType, String)> {
-    let path_components: Components<'_> = repo_path.components();
-    if check_path_components(&mut path_components.clone())
-        && check_path_string_components(ipath.clone())
-    {
-        let path_to_serve = state.repo_dir.lock().unwrap().clone()
-            + os_slash_str()
-            + &repo_path.display().to_string()
-            + "/ingredients/"
-            + ipath.as_str();
-        let file_string = match fs::read_to_string(path_to_serve) {
-            Ok(v) => v,
-            Err(e) => {
-                return status::Custom(
-                    Status::BadRequest,
-                    (
-                        ContentType::JSON,
-                        make_bad_json_data_response(
-                            format!("could not read ingredient content: {}", e).to_string(),
-                        ),
-                    ),
-                )
-            }
-        };
-        status::Custom(
-            Status::Ok,
-            (
-                ContentType::HTML,
-                format!(
-                    r#"
-                <html>
-                <head>
-                <title>Prettified</title>
-                <link rel="stylesheet" href="/webfonts/_webfonts.css">
-                </head>
-                <body>
-                <pre>
-                {}
-                </pre>
-                </body>
-                </html>
-                "#,
-                    file_string
-                ),
-            ),
-        )
-    } else {
-        status::Custom(
-            Status::BadRequest,
-            (
-                ContentType::JSON,
-                make_bad_json_data_response("bad repo path".to_string()),
-            ),
-        )
-    }
-}
-
-// CLIENTS
-
-#[get("/list-clients")]
-fn list_clients(clients: &State<Clients>) -> status::Custom<(ContentType, String)> {
-    let client_vec = public_serialize_clients(clients.lock().unwrap().clone());
-    status::Custom(
-        Status::Ok,
-        (
-            ContentType::JSON,
-            serde_json::to_string(&client_vec).unwrap(),
-        ),
-    )
-}
-
-#[get("/favicon.ico")]
-async fn serve_root_favicon() -> Redirect {
-    Redirect::to("/clients/main/favicon.ico")
-}
-
-#[get("/")]
-fn redirect_root() -> Redirect {
-    Redirect::to("/clients/main")
-}
-
-// ERROR HANDLING
-
-#[catch(404)]
-fn not_found_catcher(req: &Request<'_>) -> status::Custom<(ContentType, String)> {
-    status::Custom(
-        Status::NotFound,
-        (
-            ContentType::JSON,
-            make_bad_json_data_response(format!("Resource {} was not found", req.uri()))
-                .to_string(),
-        ),
-    )
-}
-
-#[catch(default)]
-fn default_catcher(req: &Request<'_>) -> status::Custom<(ContentType, String)> {
-    status::Custom(
-        Status::InternalServerError,
-        (
-            ContentType::JSON,
-            make_bad_json_data_response(format!("unknown error while serving {}", req.uri()))
-                .to_string(),
-        ),
-    )
-}
-
-// BUILD SERVER
+mod static_vars;
+use crate::static_vars::{DEBUG_IS_ENABLED, NET_IS_ENABLED};
 
 type MsgQueue = Arc<Mutex<VecDeque<String>>>;
 
@@ -1471,7 +437,10 @@ pub fn rocket(launch_config: Value) -> Rocket<Build> {
     }
 
     let mut my_rocket = rocket::build()
-        .register("/", catchers![not_found_catcher, default_catcher])
+        .register("/", catchers![
+            endpoints::error::not_found_catcher,
+            endpoints::error::default_catcher
+        ])
         .manage(AppSettings {
             repo_dir: Mutex::new(repo_dir_path.clone()),
             working_dir: working_dir_path.clone(),
@@ -1523,9 +492,13 @@ pub fn rocket(launch_config: Value) -> Rocket<Build> {
         })
         .mount(
             "/",
-            routes![redirect_root, serve_root_favicon, list_clients],
+            routes![
+                endpoints::clients::redirect_root,
+                endpoints::clients::serve_root_favicon,
+                endpoints::clients::list_clients
+            ],
         )
-        .mount("/notifications", routes![notifications_stream,])
+        .mount("/notifications", routes![endpoints::sse::notifications_stream,])
         .mount(
             "/settings",
             routes![
@@ -1556,32 +529,35 @@ pub fn rocket(launch_config: Value) -> Rocket<Build> {
                 endpoints::i18n::untranslated_i18n
             ],
         )
-        .mount("/navigation", routes![get_bcv, post_bcv])
+        .mount("/navigation", routes![
+            endpoints::navigation::get_bcv,
+            endpoints::navigation::post_bcv
+        ])
         .mount("/gitea", routes![
-            gitea_remote_repos,
-            get_gitea_endpoints,
-            gitea_proxy_login,
-            gitea_proxy_logout
+            endpoints::gitea::gitea_remote_repos,
+            endpoints::gitea::get_gitea_endpoints,
+            endpoints::gitea::gitea_proxy_login,
+            endpoints::gitea::gitea_proxy_logout
         ])
         .mount(
             "/git",
             routes![
-                fetch_repo,
-                list_local_repos,
-                delete_repo,
-                add_and_commit,
-                git_status
+                endpoints::git::fetch_repo,
+                endpoints::git::list_local_repos,
+                endpoints::git::delete_repo,
+                endpoints::git::add_and_commit,
+                endpoints::git::git_status
             ],
         )
         .mount(
             "/burrito",
             routes![
-                raw_ingredient,
-                get_ingredient_prettified,
-                get_ingredient_as_usj,
-                post_ingredient_as_usj,
-                raw_metadata,
-                summary_metadata
+                endpoints::burrito::raw_ingredient,
+                endpoints::burrito::get_ingredient_prettified,
+                endpoints::burrito::get_ingredient_as_usj,
+                endpoints::burrito::post_ingredient_as_usj,
+                endpoints::burrito::raw_metadata,
+                endpoints::burrito::summary_metadata
             ],
         )
         .mount("/webfonts", FileServer::from(webfonts_dir_path.clone()));
