+++ src/static_vars.rs
@@ -1,0 +1,4 @@
+use std::sync::atomic::AtomicBool;
+
+pub(crate) static NET_IS_ENABLED: AtomicBool = AtomicBool::new(false);
+pub(crate) static DEBUG_IS_ENABLED: AtomicBool = AtomicBool::new(false);
